refactored pawpals compoent
