+++ src/components/Nav/Nav.css
@@ -7,7 +7,6 @@ nav
     background-color: #363537;
     color: white;
     margin-bottom: 1.2em;
-    
 }
     
     #nav-links
@@ -23,3 +22,20 @@ nav
             cursor: pointer;
             color: gold;
         }
+    .nav-fonts
+    {
+        font-size: 1.2em;
+        font-weight: bold;
+    }
+        .nav-fonts > *
+        {
+            color: white;
+            text-decoration: none;
+        }
+
+    .nav-title
+    {
+        font-size: 1.5em;
+        color: white;
+        text-decoration: none;
+    }
