+++ src/components/PawPals/PawPal/PawPal.css
@@ -4,6 +4,7 @@
     height: 6em;
     padding: 0.5em;
     display: flex;
+
 }
 
 .pawpal-img-container
