+++ src/components/PawPals/PawPal/PawPal.js
@@ -1,8 +1,4 @@
 import React from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {Spring} from 'react-spring/renderprops';
-import {animated} from 'react-spring';
-import {get_pawpals_from_db, get_etreats_from_db} from '../../../pawbytesDB';
 
 import "./PawPal.css";
 
@@ -12,24 +8,6 @@ function PawPal(props)
     const container_bg =
     {
         backgroundColor: props.pawpal.colortheme,
-    }
-
-    const imgContainer =
-    {
-        width: "5em",
-        height:"5em",
-        float: "left"
-    }
-
-    const nameStyle =
-    {
-        float: "left",
-    }
-
-    const details =
-    {
-        marginLeft: "1.2em",
-        textTransform: "capitalize"
     }
 
     return (
